+++ packages/data/src/redux-store/thunk-middleware.ts
@@ -1,4 +1,6 @@
-export default function createThunkMiddleware( args ) {
+import type { Middleware } from 'redux';
+
+export default function createThunkMiddleware( args: unknown ): Middleware {
 	return () => ( next ) => ( action ) => {
 		if ( typeof action === 'function' ) {
 			return action( args );