+++ packages/ui/src/button/test/button.test.tsx
@@ -25,11 +25,6 @@ describe( 'Button', () => {
 
 		const onClickMock = jest.fn();
 		render(
-			// Disabling because this lint rule was meant for the
-			// `@wordpress/components` Button, but is being applied here.
-			// TODO: rework the lint rule so that it checks the package
-			// where the Button comes from.
-			// eslint-disable-next-line no-restricted-syntax
 			<Button disabled onClick={ onClickMock }>
 				Click me
 			</Button>
@@ -57,12 +52,6 @@ describe( 'Button', () => {
 
 	it( 'can be enabled explicitly when loading', () => {
 		render(
-			// Disabling because this lint rule was meant for the
-			// `@wordpress/components` Button, but is being applied here.
-			// TODO: rework the lint rule so that it checks the package
-			// where the Button comes from.
-			// TODO: Additional improvement in the original lint rule: only error if disabled=true?
-			// eslint-disable-next-line no-restricted-syntax
 			<Button loading disabled={ false }>
 				Click me
 			</Button>
@@ -76,11 +65,7 @@ describe( 'Button', () => {
 
 	it( 'supports custom render prop while retaining the default focusable when disabled behavior', () => {
 		render(
-			// Disabling because this lint rule was meant for the
-			// `@wordpress/components` Button, but is being applied here.
-			// TODO: rework the lint rule so that it checks the package
-			// where the Button comes from.
-			// eslint-disable-next-line jsx-a11y/anchor-has-content, no-restricted-syntax
+			// eslint-disable-next-line jsx-a11y/anchor-has-content
 			<Button render={ <a href="/" /> } nativeButton={ false } disabled>
 				Click me
 			</Button>